refactor(entity-currency-crypto): nest by-id accessors under legacy/
Move the by-id accessors into src/legacy/ and add FUTURE.md explaining that they are
static accessors over the compile-time registry today and become selectors over a dynamic
currency slice once the registry stops being static. Signatures are the stable seam, so call
sites are unaffected by that future change. Barrel re-export updated; no behavior change.
